feat: audio chunk generation separated by punctuation in text + multiple endpoints usable + centralisation of JSON data in Python script
Thanks to the new punctuation separation, the final generated text does not sound as choppy like before. This means that the quality of the audio has improved if there are punctuation marks in the text. In addition, further endpoints for generating the audio are now checked if one fails. The VOICE and ENDPOINT data has been transferred from the json files to the Python script to make the script easier to use.